requirements.txt
certifi==2018.4.16
chardet==3.0.4
gitdb2==2.0.3
GitPython==2.1.10
idna==2.7
in-place==0.2.0
PyYAML==3.12
requests==2.19.1
six==1.11.0
smmap2==2.0.3
urllib3==1.23
yamllint==1.11.1
PyStaticConfiguration==0.10.3
mock==2.0.0
pytest==3.6.2
sonarqube-api==1.3.1
docker==3.4.1
kazoo
pykafka
validate_email
pykwalify
